+++ test/views/templates/MainTemplateSpec.scala
@@ -46,7 +46,7 @@ class MainTemplateSpec extends SpecBase with ViewSpecAssertions with ScalaCheckP
         (content, title, lrn) =>
           val view = app.injector
             .instanceOf[MainTemplate]
-            .apply(title, timeoutEnabled = true, canSignOut = true, showBackLink = true, lrn) {
+            .apply(title, timeoutEnabled = true, showBackLink = true, lrn) {
               Html.apply(content)
             }
 
@@ -62,7 +62,7 @@ class MainTemplateSpec extends SpecBase with ViewSpecAssertions with ScalaCheckP
         (content, title, lrn) =>
           val view = app.injector
             .instanceOf[MainTemplate]
-            .apply(title, timeoutEnabled = true, canSignOut = true, showBackLink = true, lrn) {
+            .apply(title, timeoutEnabled = true, showBackLink = true, lrn) {
               Html.apply(content)
             }
 
@@ -71,6 +71,52 @@ class MainTemplateSpec extends SpecBase with ViewSpecAssertions with ScalaCheckP
           val link = getElementBySelector(doc, ".hmrc-report-technical-issue")
           getElementHref(link) mustBe s"http://localhost:9250/contact/report-technical-problem?service=CTCTraders&referrerUrl=$path"
           link.text() mustBe "Is this page not working properly? (opens in new tab)"
+      }
+    }
+
+    "when session has an auth token" - {
+      "must render timeout dialog and sign out link" in {
+        implicit lazy val request: FakeRequest[AnyContent] = FakeRequest("GET", path).withSession("authToken" -> "auth123")
+
+        forAll(Gen.alphaNumStr, Gen.alphaNumStr, arbitrary[LocalReferenceNumber]) {
+          (content, title, lrn) =>
+            val view = app.injector
+              .instanceOf[MainTemplate]
+              .apply(title, timeoutEnabled = true, showBackLink = true, lrn) {
+                Html.apply(content)
+              }
+
+            val doc = Jsoup.parse(view.toString())
+
+            val metas = getElementsByTag(doc, "meta")
+            assertElementExists(metas, _.attr("name") == "hmrc-timeout-dialog")
+            assertElementExists(metas, _.attr("data-keep-alive-url").endsWith(s"/manage-transit-movements/departures/$lrn/keep-alive"))
+
+            val link = getElementByClass(doc, "hmrc-sign-out-nav__link")
+            assertElementContainsText(link, "Sign out")
+            assertElementContainsHref(link, frontendAppConfig.signOutUrl(lrn))
+        }
+      }
+    }
+
+    "when session does not have an auth token" - {
+      "must not render timeout dialog or sign out link" in {
+        implicit lazy val request: FakeRequest[AnyContent] = FakeRequest("GET", path)
+
+        forAll(Gen.alphaNumStr, Gen.alphaNumStr, arbitrary[LocalReferenceNumber]) {
+          (content, title, lrn) =>
+            val view = app.injector
+              .instanceOf[MainTemplate]
+              .apply(title, timeoutEnabled = true, showBackLink = true, lrn) {
+                Html.apply(content)
+              }
+
+            val doc = Jsoup.parse(view.toString())
+
+            val metas = getElementsByTag(doc, "meta")
+            assertElementDoesNotExist(metas, _.attr("name") == "hmrc-timeout-dialog")
+            assertElementDoesNotExist(doc, "hmrc-sign-out-nav__link")
+        }
       }
     }
   }
@@ -87,7 +133,7 @@ class MainTemplateSpec extends SpecBase with ViewSpecAssertions with ScalaCheckP
         (content, title, lrn) =>
           val view = app.injector
             .instanceOf[MainTemplate]
-            .apply(title, timeoutEnabled = true, canSignOut = true, showBackLink = true, lrn) {
+            .apply(title, timeoutEnabled = true, showBackLink = true, lrn) {
               Html.apply(content)
             }
 
@@ -103,7 +149,7 @@ class MainTemplateSpec extends SpecBase with ViewSpecAssertions with ScalaCheckP
         (content, title, lrn) =>
           val view = app.injector
             .instanceOf[MainTemplate]
-            .apply(title, timeoutEnabled = true, canSignOut = true, showBackLink = true, lrn) {
+            .apply(title, timeoutEnabled = true, showBackLink = true, lrn) {
               Html.apply(content)
             }
 
@@ -112,6 +158,52 @@ class MainTemplateSpec extends SpecBase with ViewSpecAssertions with ScalaCheckP
           val link = getElementBySelector(doc, ".hmrc-report-technical-issue")
           getElementHref(link) must startWith(s"mailto:${config.feedbackEmail}")
           link.text() mustBe s"If you have any questions or issues, email us at ${config.feedbackEmail}"
+      }
+    }
+
+    "when session has an auth token" - {
+      "must render timeout dialog and sign out link" in {
+        implicit lazy val request: FakeRequest[AnyContent] = FakeRequest("GET", path).withSession("authToken" -> "auth123")
+
+        forAll(Gen.alphaNumStr, Gen.alphaNumStr, arbitrary[LocalReferenceNumber]) {
+          (content, title, lrn) =>
+            val view = app.injector
+              .instanceOf[MainTemplate]
+              .apply(title, timeoutEnabled = true, showBackLink = true, lrn) {
+                Html.apply(content)
+              }
+
+            val doc = Jsoup.parse(view.toString())
+
+            val metas = getElementsByTag(doc, "meta")
+            assertElementExists(metas, _.attr("name") == "hmrc-timeout-dialog")
+            assertElementExists(metas, _.attr("data-keep-alive-url").endsWith(s"/manage-transit-movements/departures/$lrn/keep-alive"))
+
+            val link = getElementByClass(doc, "hmrc-sign-out-nav__link")
+            assertElementContainsText(link, "Sign out")
+            assertElementContainsHref(link, frontendAppConfig.signOutUrl(lrn))
+        }
+      }
+    }
+
+    "when session does not have an auth token" - {
+      "must not render timeout dialog or sign out link" in {
+        implicit lazy val request: FakeRequest[AnyContent] = FakeRequest("GET", path)
+
+        forAll(Gen.alphaNumStr, Gen.alphaNumStr, arbitrary[LocalReferenceNumber]) {
+          (content, title, lrn) =>
+            val view = app.injector
+              .instanceOf[MainTemplate]
+              .apply(title, timeoutEnabled = true, showBackLink = true, lrn) {
+                Html.apply(content)
+              }
+
+            val doc = Jsoup.parse(view.toString())
+
+            val metas = getElementsByTag(doc, "meta")
+            assertElementDoesNotExist(metas, _.attr("name") == "hmrc-timeout-dialog")
+            assertElementDoesNotExist(doc, "hmrc-sign-out-nav__link")
+        }
       }
     }
   }
